Add a control signal compensation to avoid abrupt command changes
Improve the tradicional control law v(t) = A*error(t) by
adding a exponetial term that starts with the value when
the error is zero and decay with the time. This avoids
abrupt changes on the velocities command on the beginning
of the task, where the error is big. It also allow to use
higher static gain, since the soft-velocity-changing will
reduce the image motion blur.